Revise add/remove teachers and course transfer w/ teacher search
Squashed commit of the following:

commit 6cff1053454a770a5ca0b00cf2450cf77586eaa5
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Feb 23 16:47:30 2018 -0800

    Revise course transfer to allow selecting a teacher in another group

commit 7f5c92820de454ecbeedf62090a4e525f363f44e
Author: drlippman <drlippman@yahoo.com>
Date:   Thu Feb 22 20:48:49 2018 -0800

    Implement remove as coteacher

commit 4016f36ee24cbf278583b44b575ea51512d5f6bd
Author: drlippman <drlippman@yahoo.com>
Date:   Tue Feb 20 19:14:27 2018 -0800

    Replace the add/remove teacher page with new one

    This one allows searching for teachers outside one's group

+++ admin/unhidefromcourselist.php
@@ -66,7 +66,7 @@ if ($type=='take') {
 $query .= "ORDER BY ic.name";
 $stm = $DBH->prepare($query);
 $stm->execute(array(':userid'=>$userid));
-echo '<ul class="nomark">';
+echo '<ul class="nomark courselist">';
 //DB if (mysql_num_rows($result)==0) {
 if ($stm->rowCount()==0) {
 	echo '<li>No hidden courses</li>';
@@ -80,11 +80,13 @@ if ($stm->rowCount()==0) {
 			echo '<img src="../img/gears.png" alt="Options" class="mida"/></a>';
 			echo '<ul role="menu" class="dropdown-menu">';
 			echo ' <li><a href="unhidefromcourselist.php?type='.Sanitize::encodeUrlParam($type).'&tohide='.$row['id'].'">'._('Un-hide from course list').'</a></li>';
-			if ($row['ownerid']==$userid) {
+			if ($row['ownerid']==$userid && $myrights>20) {
 				echo ' <li><a href="forms.php?from=home&action=modify&id='.$row['id'].'">'._('Settings').'</a></li>';
-				echo '<li><a href="forms.php?from=home&action=chgteachers&id='.$row['id'].'">'._('Add/remove teachers').'</a></li>';
-				echo ' <li><a href="forms.php?from=home&action=transfer&id='.$row['id'].'">'._('Transfer ownership').'</a></li>';
+				echo '<li><a href="addremoveteachers.php?from=home&id='.$row['id'].'">'._('Add/remove teachers').'</a></li>';
+				echo ' <li><a href="transfercourse.php?from=home&id='.$row['id'].'">'._('Transfer ownership').'</a></li>';
 				echo ' <li><a href="forms.php?from=home&action=delete&id='.$row['id'].'">'._('Delete').'</a></li>';
+			} else if ($row['ownerid']!==$userid) {
+				echo '<li><a href="#" onclick="removeSelfAsCoteacher(this,'.$row['id'].');return false;">'._('Remove yourself as a co-teacher').'</a></li>';
 			}
 			echo '</ul></span> ';
 			echo '<a href="../course/course.php?cid='.$row['id'].'">'.Sanitize::encodeStringForDisplay($row['name']).'</a> ';
